+++ src/main/java/refinedstorage/apiimpl/storage/NBTStorage.java
@@ -43,7 +43,7 @@ public abstract class NBTStorage implements IStorage {
     /**
      * @param tag      The NBT tag we are reading from and writing the amount stored to, has to be initialized with {@link NBTStorage#createNBT()} if it doesn't exist yet
      * @param capacity The capacity of this storage, -1 for infinite capacity
-     * @param tile     A {@link TileEntity} that the NBT storage is in, will be marked dirty when storage changes
+     * @param tile     A {@link TileEntity} that the NBT storage is in, will be marked dirty when the storage changes
      */
     public NBTStorage(NBTTagCompound tag, int capacity, @Nullable TileEntity tile) {
         this.tag = tag;
@@ -236,6 +236,10 @@ public abstract class NBTStorage implements IStorage {
         tag.setInteger(NBT_PROTOCOL, PROTOCOL);
 
         return tag;
+    }
+
+    public static boolean isValid(ItemStack stack) {
+        return stack.hasTagCompound() && stack.getTagCompound().hasKey(NBT_ITEMS) && stack.getTagCompound().hasKey(NBT_STORED);
     }
 
     /**