+++ src/main/java/refinedstorage/block/BlockInterface.java
@@ -32,6 +32,11 @@ public class BlockInterface extends BlockNode {
     }
 
     @Override
+    public boolean hasConnectivityState() {
+        return true;
+    }
+
+    @Override
     public EnumPlacementType getPlacementType() {
         return null;
     }