+++ src/main/java/refinedstorage/network/MessageGridPull.java
@@ -3,33 +3,31 @@ package refinedstorage.network;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.inventory.Container;
-import net.minecraft.item.ItemStack;
-import net.minecraftforge.fml.common.network.ByteBufUtils;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import refinedstorage.api.network.IGridHandler;
 import refinedstorage.container.ContainerGrid;
 
 public class MessageGridPull extends MessageHandlerPlayerToServer<MessageGridPull> implements IMessage {
-    private ItemStack stack;
+    private int id;
     private int flags;
 
     public MessageGridPull() {
     }
 
-    public MessageGridPull(ItemStack stack, int flags) {
-        this.stack = stack;
+    public MessageGridPull(int id, int flags) {
+        this.id = id;
         this.flags = flags;
     }
 
     @Override
     public void fromBytes(ByteBuf buf) {
-        stack = ByteBufUtils.readItemStack(buf);
+        id = buf.readInt();
         flags = buf.readInt();
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
-        ByteBufUtils.writeItemStack(buf, stack);
+        buf.writeInt(id);
         buf.writeInt(flags);
     }
 
@@ -41,7 +39,7 @@ public class MessageGridPull extends MessageHandlerPlayerToServer<MessageGridPul
             IGridHandler handler = ((ContainerGrid) container).getGrid().getGridHandler();
 
             if (handler != null) {
-                handler.onExtract(message.stack, message.flags, player);
+                handler.onExtract(message.id, message.flags, player);
             }
         }
     }